chore: account for vanilla chromium

+++ build/trivalent.sh
@@ -47,8 +47,12 @@ fi
 # Fix SingletonLock if the browser isn't running
 if ! ps aux | grep "$CHROMIUM_NAME --type=zygote" | grep -v "grep" > /dev/null; then
   if [ -f "$HOME/.config/chromium/SingletonLock" ]; then
-    echo "Errm, what the sigma? This shouldn't be here."
-    rm "$HOME/.config/chromium/Singleton"*
+    if ! ps aux | grep "chromium-browser --type=zygote" | grep -v "grep" > /dev/null; then
+      echo "Errm, what the sigma? This shouldn't be here."
+      rm "$HOME/.config/chromium/Singleton"*
+    else
+      echo "Chromium is already open in this directory."
+    fi
   fi
 fi
